Move internal classes to internal package
