+++ server/netty-server/src/main/scala/sttp/tapir/server/netty/NettyOptions.scala
@@ -1,14 +1,14 @@
 package sttp.tapir.server.netty
 
 import io.netty.channel.nio.NioEventLoopGroup
-import io.netty.channel.{ChannelPipeline, EventLoopGroup}
+import io.netty.channel.{ChannelHandler, ChannelPipeline, EventLoopGroup}
 import io.netty.handler.codec.http.{HttpObjectAggregator, HttpServerCodec}
 import io.netty.handler.logging.LoggingHandler
 
 case class NettyOptions(
     eventLoopGroup: () => EventLoopGroup,
     shutdownEventLoopGroupOnClose: Boolean,
-    initPipeline: (ChannelPipeline, NettyServerHandler) => Unit
+    initPipeline: (ChannelPipeline, ChannelHandler) => Unit
 ) {
   def shutdownEventLoopGroupOnClose(shutdown: Boolean): NettyOptions = copy(shutdownEventLoopGroupOnClose = shutdown)
   def eventLoopGroup(g: EventLoopGroup): NettyOptions = copy(eventLoopGroup = () => g, shutdownEventLoopGroupOnClose = false)
@@ -24,6 +24,7 @@ object NettyOptions {
       pipeline.addLast(new HttpObjectAggregator(Integer.MAX_VALUE))
       pipeline.addLast(handler)
       pipeline.addLast(new LoggingHandler()) // TODO
+      ()
     }
   )
 }
